Extract MSIX in ostd into PCI component
