+++ kernel/comps/pci/src/arch/loongarch/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MPL-2.0
+
+pub(crate) const MSIX_DEFAULT_MSG_ADDR: u32 = 0x2ff0_0000;
+
+pub(crate) fn construct_remappable_msix_address(_remapping_index: u32) -> u32 {
+    unimplemented!()
+}
