+++ kernel/comps/pci/src/arch/x86/mod.rs
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: MPL-2.0
+
+pub(crate) const MSIX_DEFAULT_MSG_ADDR: u32 = 0xFEE0_0000;
+
+pub(crate) fn construct_remappable_msix_address(remapping_index: u32) -> u32 {
+    // Use remappable format. The bits[4:3] should be always set to 1 according to the manual.
+    let mut address = MSIX_DEFAULT_MSG_ADDR | 0b1_1000;
+
+    // Interrupt index[14:0] is on address[19:5] and interrupt index[15] is on address[2].
+    address |= (remapping_index & 0x7FFF) << 5;
+    address |= (remapping_index & 0x8000) >> 13;
+
+    address
+}
