+++ ostd/src/arch/riscv/pci.rs
@@ -74,8 +74,9 @@ pub(crate) fn init() {
     PCI_ECAM_CFG_SPACE.call_once(|| IoMem::acquire(addr_start..addr_end).unwrap());
 }
 
-pub(crate) const MSIX_DEFAULT_MSG_ADDR: u32 = 0x2400_0000;
-
-pub(crate) fn construct_remappable_msix_address(_remapping_index: u32) -> u32 {
-    unimplemented!()
+/// Encodes the bus, device, and function into an address offset in the PCI MMIO region.
+fn encode_as_address_offset(location: &PciDeviceLocation) -> u32 {
+    ((location.bus as u32) << 16)
+        | ((location.device as u32) << 11)
+        | ((location.function as u32) << 8)
 }
