+++ ostd/src/arch/x86/pci.rs
@@ -33,16 +33,3 @@ fn encode_as_port(location: &PciDeviceLocation) -> u32 {
 pub(crate) fn has_pci_bus() -> bool {
     true
 }
-
-pub(crate) const MSIX_DEFAULT_MSG_ADDR: u32 = 0xFEE0_0000;
-
-pub(crate) fn construct_remappable_msix_address(remapping_index: u32) -> u32 {
-    // Use remappable format. The bits[4:3] should be always set to 1 according to the manual.
-    let mut address = MSIX_DEFAULT_MSG_ADDR | 0b1_1000;
-
-    // Interrupt index[14:0] is on address[19:5] and interrupt index[15] is on address[2].
-    address |= (remapping_index & 0x7FFF) << 5;
-    address |= (remapping_index & 0x8000) >> 13;
-
-    address
-}
